requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
et-xmlfile==1.1.0
flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
openpyxl==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
typing-extensions==4.9.0
werkzeug==3.0.1
zipp==3.17.0
